# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
  description: VoxAgent AI Voice Bot Platform API
servers:
  - url: /api
paths:
...
  /v1/bots:
    get:
      operationId: listBots
      summary: List all registered bots
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      operationId: createBot
      summary: Register a new bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /v1/bots/{id}:
    get:
      operationId: getBot
      summary: Get bot details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found
    patch:
      operationId: updateBot
      summary: Update bot configuration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotUpdate"
      responses:
        "200":
          description: Bot updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found
    delete:
      operationId: deleteBot
      summary: Delete a bot
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    Bot:
      type: object
      required: [id, displayName, sipExtension, status, direction, supportedLanguages, defaultGreetingLanguage, timezone, endpointSilenceMs, backchannelThresholdMs, silenceRecoverySecs, activeCalls, createdAt]
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: ["string", "null"]
        sipExtension:
          type: string
        sipDomain:
          type: ["string", "null"]
        telephonyType:
          type: string
          enum: [webrtc, sip]
          default: webrtc
        whatsappNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [ONLINE, OFFLINE, BUSY, ERROR]
        activeCalls:
          type: integer
          default: 0
        direction:
          type: string
          enum: [inbound, outbound]
          default: inbound
        directionConfig:
          nullable: true
          type: object
          additionalProperties: true
        supportedLanguages:
          type: array
          items:
            type: string
          default: ["en"]
        defaultGreetingLanguage:
          type: string
          default: "en"
        timezone:
          type: string
          default: "UTC"
        endpointSilenceMs:
          type: integer
          default: 1200
        backchannelThresholdMs:
          type: integer
          default: 700
        silenceRecoverySecs:
          type: integer
          default: 6
        llmChainJson:
          type: ["array", "null"]
          items:
            type: object
            additionalProperties: true
        sttMapJson:
          type: ["object", "null"]
          additionalProperties: true
        ttsMapJson:
          type: ["object", "null"]
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    BotInput:
      type: object
      required: [displayName, sipExtension]
      properties:
        displayName:
          type: string
        email:
          type: ["string", "null"]
        sipExtension:
          type: string
        sipDomain:
          type: ["string", "null"]
        telephonyType:
          type: string
          enum: [webrtc, sip]
          default: webrtc
        whatsappNumber:
          type: ["string", "null"]
        direction:
          type: string
          enum: [inbound, outbound]
        directionConfig:
          type: object
          additionalProperties: true
        supportedLanguages:
          type: array
          items:
            type: string
        defaultGreetingLanguage:
          type: string
        timezone:
          type: string
        endpointSilenceMs:
          type: integer
        backchannelThresholdMs:
          type: integer
        silenceRecoverySecs:
          type: integer
        llmChainJson:
          type: ["array", "null"]
          items:
            type: object
            additionalProperties: true
        sttMapJson:
          type: ["object", "null"]
          additionalProperties: true
        ttsMapJson:
          type: ["object", "null"]
          additionalProperties: true
    BotUpdate:
      type: object
      properties:
        displayName:
          type: string
        email:
          type: ["string", "null"]
        sipExtension:
          type: string
        sipDomain:
          type: ["string", "null"]
        telephonyType:
          type: string
          enum: [webrtc, sip]
        whatsappNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [ONLINE, OFFLINE, BUSY, ERROR]
        direction:
          type: string
          enum: [inbound, outbound]
        directionConfig:
          type: object
          additionalProperties: true
        supportedLanguages:
          type: array
          items:
            type: string
        defaultGreetingLanguage:
          type: string
        timezone:
          type: string
        endpointSilenceMs:
          type: integer
        backchannelThresholdMs:
          type: integer
        silenceRecoverySecs:
          type: integer
        llmChainJson:
          type: ["array", "null"]
          items:
            type: object
            additionalProperties: true
        sttMapJson:
          type: ["object", "null"]
          additionalProperties: true
        ttsMapJson:
          type: ["object", "null"]
          additionalProperties: true
...